Add new G33 list output: UBL-Like.
Added a new output to G33 L. Using L1 now outputs our mesh data in UBL-Like format.
This was specifically added so that we can visualize the data on sites such as http://lokspace.eu/3d-printer-auto-bed-leveling-mesh-visualizer/

At the end of the output there's a hint with a regex that'll strip the timestamps from repetier-server's console after being copied into a text editor.
Use it in a text editor such as Notepad++.

TODO: importing a mesh fails when we have an empty/fresh eeprom.