Added config feature NUKE_EEPROM_ON_COMPILE.
Small feature to erase and reset the previously stored eeprom on the printer after a firmware flash, but only after a flash. This is done via the use of a compile-time timestamp that'll now be saved on the eeprom.

This doesn't require any RTC or external libraries/includes, and works automatically. 

Mostly aimed towards debugging and development.