Fix G33 X Y Z Bump correction importing.
The set function now quickly checks if we have our bed rectangle properly set, and sets it if it's not. This let's us import bumps into the EEPROM again.

To export a bump map:
- Run a G33 L0 command in your console. (Repetier-Server, etc)
- Copy paste the data into notepad++ or similar
- Do a find and replace in notepad using the regex (.*:\s+) to remove all the timestamps and leave us with raw G33 commands we can use later.

To import a bump map:
- Copy your raw G33 commands and paste them into a repetier-server quick command, or even an empty gcode file.
- Run it.
As long as your bed minimums and maximums are the same as when you created the original bump map, the import should go smoothly.