Z-Probe check if probe never triggered and fix secureDistance
I think this was the original intention of secureDistance (?) and I believe this still gives us correct distances for printers that home at z max, since we move to the optimum probing height in LevelingMethod before a probe run.

I added a threshold check if we have remaining steps, which assumes a bed crash since in theory we should never reach that close. 
(Could just check for zero steps, though we might have a odd trigger in the middle of a crash etc.)

In adding this I think I stumbled across a small bug in stepsRemaining:

Rarely stepsRemaining was still actually set to something (previous bed move iirc) before moveByPrinter, even after waitForEndOfMoves. This would also affect the first z probe height since it'd give us the wrong steps.

I clear it out before we move in case it is, and it does consistently give us the right remaining steps value now.