WIP: IO_SENSORS concept (Mixed sensors)
I wanted to add a DHT22 sensor to measure humidity inside my printer's enclosure so created this concept of mixed sensors that can report various data. 
I'M NOT entirely sure if this is a good route as "sensors" is quite vague, we already have some sensors in IO_TEMPERATURE, and the only other data these return is humidity. Perhaps it should just be IO_HUMIDITY. 

Anyway, this only works for interrupt capable pins at the moment. defining a main/common humidity sensor in your config appends a small add-on print to autoreported/M105 commands that shows RH and it's temperature. That's it for now.