isNativeUSB - re-implement using GCodeSource classes for modularity
Updated this to use this new structure as well as keep track of which serial source is the usb host source. Helps build the path towards more usb-host specific functionality 

Switched back to using a pointer compare to the serialUSB class in the end anyways rather than the strncmp. (since it doesn't work here)
I doubt there should be too much trouble adapting even if the class is renamed/doesn't exist in the future on a different HAL.